+++ src/decl/type-info-decl.h
@@ -1,8 +1,2 @@
 static
-r_obj* syms_vec_is_vector_dispatch;
-
-static
-r_obj* fns_vec_is_vector_dispatch;
-
-static
 enum vctrs_type vec_base_typeof(r_obj* x, bool proxied);