+++ src/subscript.c
@@ -31,7 +31,7 @@ r_obj* vec_as_subscript_opts(r_obj* subscript,
   }
   KEEP_AT(subscript, subscript_pi);
 
-  if (!vec_is_vector(subscript)) {
+  if (!obj_is_vector(subscript)) {
     *err = new_error_subscript_type(subscript, opts, r_null);
     FREE(2);
     return r_null;