+++ src/type-info.c
@@ -112,11 +112,11 @@ bool vec_is_list(r_obj* x) {
   return (type == VCTRS_CLASS_list) || (type == VCTRS_CLASS_bare_asis);
 }
 
-r_obj* ffi_vec_is_vector(r_obj* x) {
-  return r_lgl(vec_is_vector(x));
+r_obj* ffi_obj_is_vector(r_obj* x) {
+  return r_lgl(obj_is_vector(x));
 }
 
-bool vec_is_vector(r_obj* x) {
+bool obj_is_vector(r_obj* x) {
   if (x == r_null) {
     return false;
   }
@@ -135,7 +135,7 @@ bool list_all_vectors(r_obj* x) {
   if (r_typeof(x) != R_TYPE_list) {
     r_stop_unexpected_type(r_typeof(x));
   }
-  return r_list_all_of(x, &vec_is_vector);
+  return r_list_all_of(x, &obj_is_vector);
 }
 
 
@@ -196,12 +196,5 @@ const char* vec_type_as_str(enum vctrs_type type) {
 
 
 void vctrs_init_type_info(r_obj* ns) {
-  syms_vec_is_vector_dispatch = r_sym("vec_is_vector");
-  fns_vec_is_vector_dispatch = r_eval(syms_vec_is_vector_dispatch, ns);
+
 }
-
-static
-r_obj* syms_vec_is_vector_dispatch = NULL;
-
-static
-r_obj* fns_vec_is_vector_dispatch = NULL;