+++ NEWS.md
@@ -1,6 +1,6 @@
 # vctrs (development version)
 
-* New `vec_is_vector()`, `vec_check_vector()`, and `vec_check_size()` validation
+* New `obj_is_vector()`, `vec_check_vector()`, and `vec_check_size()` validation
   helpers. We believe these are a better approach to vector validation than
   `vec_assert()` and `vec_is()`, which have been marked as questioning because
   the semantics of their `ptype` arguments are hard to define and can often be